Support rendering of embedded forms via the HTML compiler
- Move the current HTML compiler/renderer into a new `RichText::Rendering` namespace
- Add an `EmbeddedFormRenderer` that is initialised with a container of callable objects for rendering embedded forms based on their name
- Pass the embedded form renderer to the HTML compiler and call it whenever a “formalist” entity is found in the rich text AST
